requirements.txt
appdirs
aria2p
beautifulsoup4
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
js2py
lxml
psutil
Pyrogram
python-dotenv
python-magic
python-telegram-cheems
requests
telegraph
tenacity
TgCrypto
youtube_dl
megasdkrestclient>=0.1.1,<1.0.0